requirements.txt
# Core dependencies (used in ShapeGeometryEncoder, ShapePoseEncoder, ZernikeBasisCorpus, XShapeCorpus, utils)
numpy==1.26.4
scipy==1.12.0
matplotlib==3.9.4
shapely==2.0.7
PyYAML==6.0.2
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
torchinfo==1.8.0
scikit-learn==1.6.1
scikit-image==0.24.0
pandas==1.5.3
Pillow==11.3.0
opencv-python==4.11.0.86
tqdm==4.67.3
pyefd==1.7.0
triangle==20250106
librosa==0.11.0
pesq
pyroomacoustics
pystoi
geopandas